requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
emoji==2.9.0
idna==3.6
pillow==10.2.0
pilmoji==2.0.4
requests==2.31.0
urllib3==2.1.0
